+++ mall-product/src/main/java/com/katzenyasax/mall/product/entity/SkuImagesEntity.java
@@ -39,6 +39,6 @@ public class SkuImagesEntity implements Serializable {
 	/**
 	 * 默认图[0 - 不是默认图，1 - 是默认图]
 	 */
-	private Integer defaultImg;
+	private Long defaultImg;
 
 }
